requirements.txt
attrs==24.3.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.6.1
exceptiongroup==1.2.2
flask==3.0.3
gunicorn==23.0.0
h11==0.14.0
idna==3.10
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.5
MarkupSafe==2.1.5
outcome==1.3.0.post0
packaging==24.2
pymongo==4.10.1
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
selenium==4.27.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
trio==0.27.0
trio-websocket==0.11.1
typing-extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
werkzeug==3.0.6
wsproto==1.2.0
zipp==3.20.2
